Removed redundant styles and normalized icon size.
I refactored the styles to remove duplication of definitions. Also
manually defined the icon size regardless of the setting chosen for the
tool-bar package. This should eventually be configured via a UI scaling
option in the material-ui settings.

I seem to have broken the toolbar when in the top position so that will
need to be looked at at some point and fixed.

The hover color of the buttons also needs to be defined correctly. I
wasn't sure whether it should be defined here or if there was a global
definition for button styles somewhere which would be more appropriate.